Move video to Atoms
